requirements.txt
python-slugify==1.2.6
hdx-python-api==3.4.5
-r docker-requirements.txt